# src/docs/api-docs.json
{
  "openapi" : "3.0.0",
  "info" : {
    "version" : "1.0.1",
    "title" : "Lobster",
    "description" : "Go to /login/ to obtain your token. Authorize with the following value: 'Token {your_token}'"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "ApiKeyAuth" : [ ]
  } ],
  "components" : {
    "securitySchemes" : {
      "ApiKeyAuth" : {
        "type" : "apiKey",
        "in" : "header",
        "name" : "Authorization"
      }
    },
    "schemas" : {
...
      "Course" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "readOnly" : true
          },
          "short_name" : {
            "type" : "string"
          },
          "full_name" : {
            "type" : "string"
          },
          "term" : {
            "type" : "string",
            "enum" : [ "fall", "winter", "spring", "summer" ]
          },
          "year" : {
            "type" : "integer"
          }
        }
      },
      "Project" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "readOnly" : true
          },
          "exercise_id" : {
            "type" : "integer"
          },
          "last_modified" : {
            "type" : "string",
            "readOnly" : true
          },
          "contents" : {
            "type" : "object"
          }
        }
      },
...
    }
  },
  "paths" : {
    "/public/projects/{project_id}/" : {
      "get" : {
        "tags" : [ "Public" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "project_id",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "summary" : "Gets details on a project on lobster",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Project"
                }
              }
            }
          },
          "404" : {
            "description" : "NOT FOUND"
          }
        }
      }
    },
...
    "/public/courses/{course_id}/" : {
      "get" : {
        "tags" : [ "Public" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "course_id",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "summary" : "Gets details of a course",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Course"
                }
              }
            }
          },
          "404" : {
            "description" : "NOT FOUND"
          }
        }
      }
    },
    "/public/courses/{short_name}/{term}/{year}" : {
      "get" : {
        "tags" : [ "Public" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "short_name",
          "schema" : {
            "type" : "string"
          },
          "required" : true
        }, {
          "in" : "path",
          "name" : "term",
          "schema" : {
            "type" : "string",
            "enum" : [ "fall", "winter", "spring", "summer" ]
          },
          "required" : true
        }, {
          "in" : "path",
          "name" : "year",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "summary" : "Gets details of a course",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Course"
                }
              }
            }
          },
          "404" : {
            "description" : "NOT FOUND"
          }
        }
      }
    },
    "/courses/" : {
      "get" : {
        "tags" : [ "Courses" ],
        "summary" : "Gets a list of courses on lobster",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "courses" : {
                      "type" : "array",
                      "items" : {
                        "$ref" : "#/components/schemas/Course"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Courses" ],
        "summary" : "Adds a course to lobster",
        "requestBody" : {
          "description" : "Course Post Object",
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Course"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Course"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{short_name}/{term}/{year}/" : {
      "get" : {
        "tags" : [ "Courses" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "short_name",
          "schema" : {
            "type" : "string"
          },
          "required" : true
        }, {
          "in" : "path",
          "name" : "term",
          "schema" : {
            "type" : "string",
            "enum" : [ "fall", "winter", "spring", "summer" ]
          },
          "required" : true
        }, {
          "in" : "path",
          "name" : "year",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "summary" : "Gets details on a course on lobster",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Course"
                }
              }
            }
          }
        }
      }
    },
    "/courses/{course_id}/" : {
      "get" : {
        "tags" : [ "Courses" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "course_id",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "summary" : "Gets details on a course on lobster",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Course"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "Courses" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "course_id",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "requestBody" : {
          "description" : "Course Post Object",
          "required" : false,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Course"
              }
            }
          }
        },
        "summary" : "Edits a course on lobster",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Course"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Courses" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "course_id",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "summary" : "Deletes a course on lobster",
        "responses" : {
          "204" : {
            "description" : "OK"
          }
        }
      }
    },
    "/courses/{course_id}/copy/" : {
      "post" : {
        "tags" : [ "Courses" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "course_id",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "summary" : "Copies an existing course on lobster. Copies exercises and course admins but does not copy other users",
        "responses" : {
          "201" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Course"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/" : {
      "post" : {
        "tags" : [ "Projects" ],
        "summary" : "Creates a new project",
        "requestBody" : {
          "description" : "Project data",
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Project"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/" : {
      "get" : {
        "tags" : [ "Projects" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "project_id",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "summary" : "Gets details on a project on lobster",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Project"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Projects" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "project_id",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "summary" : "Deletes a project on lobster",
        "responses" : {
          "204" : {
            "description" : "OK"
          }
        }
      },
      "patch" : {
        "tags" : [ "Projects" ],
        "parameters" : [ {
          "in" : "path",
          "name" : "project_id",
          "schema" : {
            "type" : "integer"
          },
          "required" : true
        } ],
        "summary" : "Modifies a project on lobster",
        "requestBody" : {
          "description" : "Project data",
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Project"
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "description" : "OK"
          }
        }
      }
    },
...
  }
}